pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdaf-gh-actions"
version = "0.1.0"
description = "A tool for automating the setup of SAP Deployment Automation Framework on Azure"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyGithub>=1.55",
    "textual>=0.13.0",
    "requests>=2.27.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.1.0",
    "isort>=5.10.0",
    "mypy>=0.931",
    "flake8>=4.0.1"
]

[project.scripts]
sdaf-gh-actions = "sdaf-gh-actions.main:main"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
